requirements.txt
fastapi[standard]
sqlalchemy
psycopg
psycopg-binary
python-dotenv
httpx
asyncio
pydantic
pyjwt
passlib[bcrypt]

